fix(cmd): relocate temp files across filesystem boundaries
